% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.equa_profile}
\alias{plot.equa_profile}
\title{Plot EQuA profiles}
\usage{
\method{plot}{equa_profile}(
  x,
  type = c("qea", "qeaa"),
  rows = seq_len(min(6L, nrow(x$qea))),
  xlab = "Percentile level",
  ylab = NULL,
  ...
)
}
\arguments{
\item{x}{An \code{equa_profile} or \code{equa_crossfit} object.}

\item{type}{Either \code{"qea"} or \code{"qeaa"}.}

\item{rows}{Integer row indices to display.}

\item{xlab, ylab}{Axis labels.}

\item{...}{Additional arguments passed to \code{\link[graphics:matplot]{graphics::matplot()}}.}
}
\value{
The plotted object, invisibly.
}
\description{
Plot EQuA profiles
}
