% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equa-package.R
\docType{package}
\name{EQuA-package}
\alias{EQuA}
\alias{EQuA-package}
\title{EQuA: Post-Hoc Quantile Calibration of Age Estimators}
\description{
EQuA transforms an existing scalar age estimate into individual
quantile-age and quantile-age-acceleration profiles. The development
package implements Po-EQuA, End-EQuA, and Ex-EQuA, prediction,
support diagnostics, and repeated cross-fitting in pure R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Menghan898/EQuA}
  \item Report bugs at \url{https://github.com/Menghan898/EQuA/issues}
}

}
\author{
\strong{Maintainer}: Menghan Yi \email{menghany@umich.edu}

}
\keyword{internal}
